+++ src/pytorch_metric_learning/losses/normalized_softmax_loss.py
@@ -11,7 +11,7 @@ class NormalizedSoftmaxLoss(WeightRegularizerMixin, BaseMetricLossFunction):
         self.cross_entropy = torch.nn.CrossEntropyLoss(reduction='none')
         
     def compute_loss(self, embeddings, labels, indices_tuple):
-        miner_weights = lmu.convert_to_weights(indices_tuple, labels)
+        miner_weights = lmu.convert_to_weights(indices_tuple, labels, dtype=embeddings.dtype)
         normalized_W = torch.nn.functional.normalize(self.W, p=2, dim=0)
         exponent = torch.matmul(embeddings, normalized_W) / self.temperature
         unweighted_loss = self.cross_entropy(exponent, labels)
